+++ src/discovery/index.js
@@ -1,0 +1,14 @@
+export {
+  collectSearchConsoleObservations,
+  compareSearchDiscoveryObservations,
+  pagesForQuery,
+  queriesForPage,
+  relationshipObservations,
+} from "./searchConsole.js";
+
+export {
+  createPassiveBacklinkObservation,
+  compareBacklinkObservations,
+  backlinksForTarget,
+  referringDomainsForTarget,
+} from "./backlinks.js";