+++ src/representation/index.js
@@ -1,0 +1,31 @@
+export {
+  createPrompt,
+  createPromptSet,
+  verifyPromptSet,
+  promptReference,
+  comparePromptSets,
+} from "./prompts.js";
+
+export {
+  createRepresentationObservation,
+  collectRepresentationObservations,
+  compareRepresentationObservations,
+} from "./representation.js";
+
+export {
+  createResponseSpan,
+  createCitationObservation,
+  createSourceRetrievalObservation,
+  verifyCitationProvenance,
+  deriveRepresentationPrevalence,
+} from "./provenance.js";
+
+export {
+  resolveObservedEntity,
+  createCompetitorRepresentationObservation,
+  verifyCompetitorRepresentation,
+} from "./competitors.js";
+
+export {
+  verifyContentEvidenceBoundary,
+} from "./contentBoundary.js";